fix: 32处bug修复 + README中英双语重写
严重修复:
- ETC2 RGBA8 alpha通道被DecodeEtc2ColorBlock覆盖
- TextureDecoder width*height*4 int溢出 (9处)
- DXT.DecompressImage 溢出+越界
- ImageToTexConverter 像素数据分配溢出
- PackageEntry.Offset/Length int→long
- Package.HeaderSize int→long
- PackageReader dataStart int截断→long
- PackageWriter currentOffset int溢出→long

高优先级修复:
- Extract Path.GetDirectoryName null崩溃
- Extract dynamic JSON无null检查
- Extract Split未过滤空字符串
- Extract Magic比较未用OrdinalIgnoreCase
- Extract非递归模式跳过根目录pkg
- Replace FirstImage.FirstMipmap空引用链
- Replace ConvertToTexBytes未检查null
- Convert扩展名比较未用OrdinalIgnoreCase
- Repack Substring无边界保护
- TexToImageConverter Image未dispose (ConvertToImage+ConvertToGif)
- TexImageReader byteCount负数绕过检查
- TexMipmapDecompressor LZ4不验证解压结果
- TexFrameInfoContainerReader Frames[0]越界
- Info.cs适配long类型

低优先级修复:
- Extensions.ReadNString无null终止符异常
- NumberIsPowerOfTwo负数无限循环
- RG88.Equals类型不匹配
- PackageWriter NullReferenceException→ArgumentException
- Convert.cs多余?.去除

其他:
- README.md中英双语重写
- 全部Console.WriteLine中文化
- T()双语支持 (--en)

+++ RePKG.Core/Package/Package.cs
@@ -5,7 +5,7 @@ namespace RePKG.Core.Package
     public class Package
     {
         public string Magic { get; set; }
-        public int HeaderSize { get; set; }
+        public long HeaderSize { get; set; }
 
         public List<PackageEntry> Entries { get; } = new List<PackageEntry>();
     }
